Crate clock_ticks

Functions

precise_time_ns
Returns the current value of a high-resolution performance counter in nanoseconds since an unspecified epoch.

precise_time_s
Returns the current value of a high-resolution performance counter in seconds since an unspecified epoch.